requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django
django-ckeditor==6.4.2
django-debug-toolbar==3.5.0
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.3
Faker==13.15.0
httpie==3.2.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pymemcache==3.5.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
celery
flower
gunicorn